requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
twilio==8.8.0
cryptography==41.0.7
mysql-connector-python==8.1.0
